scripts/requirements.txt
accelerate>=0.25.0
transformers>=4.36.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pyyaml>=5.3
python-frontmatter
markdown
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm
python-dotenv
requests>=2.31.0
toml>=0.10.2
flowhunt~=3.0
nltk>=3.8
numpy>=1.21.0
scikit-learn>=1.0.0
hdbscan>=0.8.33
Pillow>=10.0.0
yake>=0.4.8
tomli_w>=1.0.0